+++ .env.example
@@ -3,6 +3,6 @@ FRONTEND_DEV_PORT=2222
 STORYBOOK_DEV_PORT=3333
 # BACKEND_DEV_PORT=4444
 
-# local test ports for e2e testing and preview
-FRONTEND_TEST_PORT=5555
+# for e2e testing and production preview
+FRONTEND_URL=http://localhost:5555
 # BACKEND_TEST_PORT=6666
