+++ docker-compose.e2e.yaml
@@ -8,7 +8,7 @@ services:
       - '${FRONTEND_TEST_PORT:-5000}:${FRONTEND_TEST_PORT:-5000}'
     volumes:
       - ./frontend:/app/frontend
-      - /app/node_modules
+      - /app/frontend/node_modules
     command: sh -c "npm run preview -- --port ${FRONTEND_PROD_PORT:-5000} --host"
     healthcheck:
       test: curl -f http://localhost:${FRONTEND_PROD_PORT:-5000} || exit 1
@@ -26,12 +26,12 @@ services:
       frontend-preview:
         condition: service_healthy
     environment:
-      - FRONTEND_URL=http://localhost:${FRONTEND_TEST_PORT:-5000}
+      - FRONTEND_URL=${FRONTEND_URL:-http://localhost:5000}
       - CI=${CI}
       - ARGOS_TOKEN=${ARGOS_TOKEN}
     volumes:
       - ./e2e:/app/e2e
-      - /app/node_modules
+      - /app/e2e/node_modules
     command: npm run test
     network_mode: host
     # https://playwright.dev/docs/docker#recommended-docker-configuration
