+++ docker-compose.yaml
@@ -4,8 +4,8 @@ services:
     ports:
       - '${FRONTEND_DEV_PORT:-3000}:${FRONTEND_DEV_PORT:-3000}'
     volumes:
-      - ./frontend:/app
-      - /app/node_modules
+      - ./frontend:/app/frontend
+      - /app/frontend/node_modules
     command: npm run dev -- --port ${FRONTEND_DEV_PORT:-3000} --host
 
   storybook-dev:
@@ -13,6 +13,6 @@ services:
     ports:
       - '${STORYBOOK_DEV_PORT:-4000}:${STORYBOOK_DEV_PORT:-4000}'
     volumes:
-      - ./frontend:/app
-      - /app/node_modules
+      - ./frontend:/app/frontend
+      - /app/frontend/node_modules
     command: npm run storybook -- --port ${STORYBOOK_DEV_PORT:-4000} --no-open
