+++ frontend/src/tests/basic.test.ts
@@ -1,0 +1,5 @@
+import { expect, test } from 'vitest'
+
+test('Basic Test', () => {
+  expect(true).toBe(true)
+})
