+++ package.json
@@ -14,8 +14,11 @@
     "fix:format": "prettier --write .",
     "fix:lint": "eslint --fix",
     "fix": "npm run fix:format && npm run fix:lint",
-    "test:e2e": "docker compose -f docker-compose.e2e.yaml up --build --abort-on-container-exit",
-    "test": "npm run test:e2e",
+    "test:frontend": "npm run --workspace=frontend test",
+    "test:e2e": "npm run --workspace=e2e test",
+    "test:e2e-local": "docker compose -f docker-compose.e2e.yaml up --build --abort-on-container-exit",
+    "test": "npm run test:frontend && npm run test:e2e-local",
+    "build:frontend": "npm run --workspace=frontend build",
     "sync-lockfile:frontend": "docker compose -f docker-compose.e2e.yaml run --rm frontend-preview npm install",
     "sync-lockfile:e2e": "docker compose -f docker-compose.e2e.yaml run --rm --no-deps e2e-tests npm install",
     "sync-lockfiles": "npm run sync-lockfile:frontend && npm run sync-lockfile:e2e"
